Break up and rename lsystem code.
