+++ lmorpho/CMakeLists.txt
@@ -1,0 +1,5 @@
+add_executable(lmorpho lmorpho.cpp lsystem.cpp)
+
+# target_link_libraries(miniapp.exe LINK_PUBLIC nestmc)
+# target_link_libraries(miniapp.exe LINK_PUBLIC ${EXTERNAL_LIBRARIES})
+
